backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.9.0
yfinance>=0.2.45
pandas>=2.2.0
pandas-ta>=0.4.67b0
numpy>=2.2.0
aiofiles>=24.1.0
python-multipart>=0.0.12

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# WebSocket
websockets>=12.0

# Environment
python-dotenv>=1.0.0
